Add files via upload
LionLamb Phonic-Sounds-Story-Box is an app that allows you generate African children  stories using any of the 42 English language sounds  or  an idea in expressed in one sentence.  It generates the story image, as well as the as the audio.  The story can be copied and the image downloaded.  
Brought To You By: A-Digital-Arena Academy.